refactor(frontend/src/shared/ui/Form/Form.jsx):
move Form component to shared/ui/Form

for better UI organization
